+++ CHANGELOG.md
@@ -18,6 +18,7 @@ See [Attachments] wiki page for details.
 - Add events for user creation/update (@glensc, #289)
 - Harden closeIssue access checks/validation in RPC code (@glensc, #287)
 - Drop table prefix uses from code (@glensc, #296)
+- Handle PDOException HY000 non-numeric values (@glensc, #301)
 
 [3.3.0]: https://github.com/eventum/eventum/compare/v3.2.3...v3.3.0
 [Attachments]: https://github.com/eventum/eventum/wiki/Attachments